@@ -2,13 +2,14 @@ package com.pengxh.app.weatherplus.adapter;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.pengxh.app.weatherplus.R;
 import com.pengxh.app.weatherplus.bean.NetWeatherBean;
@@ -18,12 +19,10 @@ import java.util.List;
 
 public class HourlyRecyclerViewAdapter extends RecyclerView.Adapter {
 
-    private Context mContext;
     private List<NetWeatherBean.ResultBeanX.ResultBean.HourlyBean> hourlyBeanList;
     private LayoutInflater inflater;
 
     public HourlyRecyclerViewAdapter(Context mContext, List<NetWeatherBean.ResultBeanX.ResultBean.HourlyBean> hourlyBeanList) {
-        this.mContext = mContext;
         this.hourlyBeanList = hourlyBeanList;
         inflater = LayoutInflater.from(mContext);
     }
@@ -45,7 +44,7 @@ public class HourlyRecyclerViewAdapter extends RecyclerView.Adapter {
         itemHolder.bindHolder(hourlyBeanList.get(position));
     }
 
-    class HourlyRecyclerViewHolder extends RecyclerView.ViewHolder {
+    static class HourlyRecyclerViewHolder extends RecyclerView.ViewHolder {
 
         private TextView mRecyclerView_hourly_time;
         private TextView mRecyclerView_hourly_weather;
