@@ -1,5 +1,6 @@
 package com.pengxh.app.weatherplus.ui;
 
+import android.annotation.SuppressLint;
 import android.content.Intent;
 import android.graphics.Color;
 import android.graphics.drawable.ColorDrawable;
@@ -11,6 +12,7 @@ import android.widget.TextView;
 import com.gyf.immersionbar.ImmersionBar;
 import com.pengxh.app.multilib.base.BaseNormalActivity;
 import com.pengxh.app.multilib.utils.DensityUtil;
+import com.pengxh.app.multilib.utils.SaveKeyValues;
 import com.pengxh.app.multilib.widget.EasyToast;
 import com.pengxh.app.multilib.widget.swipemenu.SwipeMenu;
 import com.pengxh.app.multilib.widget.swipemenu.SwipeMenuCreator;
@@ -19,14 +21,9 @@ import com.pengxh.app.multilib.widget.swipemenu.SwipeMenuListView;
 import com.pengxh.app.weatherplus.R;
 import com.pengxh.app.weatherplus.adapter.CityListAdapter;
 import com.pengxh.app.weatherplus.bean.CityListWeatherBean;
-import com.pengxh.app.weatherplus.event.TagEvent;
 import com.pengxh.app.weatherplus.utils.OtherUtil;
 import com.pengxh.app.weatherplus.utils.SQLiteUtil;
-import com.pengxh.app.weatherplus.utils.SaveKeyValues;
 
-import org.greenrobot.eventbus.EventBus;
-import org.greenrobot.eventbus.Subscribe;
-import org.greenrobot.eventbus.ThreadMode;
 import org.json.JSONException;
 import org.json.JSONObject;
 
@@ -63,18 +60,16 @@ public class CityListActivity extends BaseNormalActivity implements View.OnClick
     private SQLiteUtil sqLiteUtil;
 
     @Override
-    public void initView() {
-        setContentView(R.layout.activity_citylist);
-        ImmersionBar.with(this)
-                .statusBarColor("#00BAFF")
-                .fitsSystemWindows(true)
-                .init();
+    public int initLayoutView() {
+        return R.layout.activity_citylist;
     }
 
+    @SuppressLint("SetTextI18n")
     @Override
     public void initData() {
+        ImmersionBar.with(this).statusBarColor("#00BAFF").fitsSystemWindows(true).init();
         //设置第一个item的城市天气。后续改为实时更新的效果
-        String weatherJson = (String) SaveKeyValues.getValue("location_weather", "weatherMap", "");
+        String weatherJson = (String) SaveKeyValues.getValue("weatherMap", "");
         if (!weatherJson.equals("")) {
             /**
              * {"templow":"19","img":"301","color":"#FFFF00","city":"西城区","weather":"雨","quality":"良","temphigh":"28"}
@@ -121,7 +116,7 @@ public class CityListActivity extends BaseNormalActivity implements View.OnClick
                 switch (index) {
                     case 0:
                         //先删除数据库数据，再删除List，不然会出现角标越界
-                        sqLiteUtil.deleteCityByName(listWeatherBeans.get(position).getCityName());
+                        sqLiteUtil.deleteCityByName(listWeatherBeans.get(position).getCity());
                         listWeatherBeans.remove(position);
                         cityAdapter.notifyDataSetChanged();
                         //TODO 发送del广播
@@ -166,37 +161,37 @@ public class CityListActivity extends BaseNormalActivity implements View.OnClick
                 break;
             case R.id.mImageView_title_add:
                 Intent intent = new Intent(this, SelectCityActivity.class);
-                intent.putExtra("district", (String) SaveKeyValues.getValue("location", "district", ""));
+                intent.putExtra("district", (String) SaveKeyValues.getValue("district", ""));
                 startActivity(intent);
                 break;
         }
     }
 
-    @Override
-    protected void onStart() {
-        super.onStart();
-        EventBus.getDefault().register(this);
-    }
-
-    @Override
-    protected void onStop() {
-        super.onStop();
-        EventBus.getDefault().unregister(this);
-    }
-
-    @Subscribe(sticky = true, threadMode = ThreadMode.MAIN)
-    public void onEventMainThread(TagEvent event) {
-        String className = event.getClassName();
-        int msg = event.getMsg();
-        if (SelectCityActivity.class.getSimpleName().equals(className) && msg == 1) {
-            //刷新UI
-            sqLiteUtil = SQLiteUtil.getInstance();
-            listWeatherBeans = sqLiteUtil.loadCityList();
-            if (listWeatherBeans.size() > 0) {
-                cityAdapter = new CityListAdapter(this, listWeatherBeans);
-                mSwipeMenuListView.setAdapter(cityAdapter);
-            }
-        }
-        EventBus.getDefault().removeStickyEvent(event);
-    }
+//    @Override
+//    protected void onStart() {
+//        super.onStart();
+//        EventBus.getDefault().register(this);
+//    }
+//
+//    @Override
+//    protected void onStop() {
+//        super.onStop();
+//        EventBus.getDefault().unregister(this);
+//    }
+//
+//    @Subscribe(sticky = true, threadMode = ThreadMode.MAIN)
+//    public void onEventMainThread(TagEvent event) {
+//        String className = event.getClassName();
+//        int msg = event.getMsg();
+//        if (SelectCityActivity.class.getSimpleName().equals(className) && msg == 1) {
+//            //刷新UI
+//            sqLiteUtil = SQLiteUtil.getInstance();
+//            listWeatherBeans = sqLiteUtil.loadCityList();
+//            if (listWeatherBeans.size() > 0) {
+//                cityAdapter = new CityListAdapter(this, listWeatherBeans);
+//                mSwipeMenuListView.setAdapter(cityAdapter);
+//            }
+//        }
+//        EventBus.getDefault().removeStickyEvent(event);
+//    }
 }
