@@ -1,62 +1,95 @@
 package com.pengxh.app.weatherplus.ui;
 
-import android.Manifest;
-import android.annotation.SuppressLint;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
-import android.content.IntentFilter;
-import android.content.pm.PackageManager;
-import android.os.Handler;
-import android.os.Message;
-import android.support.v4.app.Fragment;
-import android.support.v4.content.ContextCompat;
-import android.support.v4.view.ViewPager;
 import android.util.Log;
+import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 
-import com.aihook.alertview.library.AlertView;
-import com.aihook.alertview.library.OnItemClickListener;
+import androidx.viewpager.widget.ViewPager;
+
+import com.gyf.immersionbar.ImmersionBar;
 import com.pengxh.app.multilib.base.BaseNormalActivity;
+import com.pengxh.app.multilib.utils.DensityUtil;
+import com.pengxh.app.multilib.utils.SaveKeyValues;
 import com.pengxh.app.weatherplus.R;
-import com.pengxh.app.weatherplus.adapter.WeatherPageAdapter;
 import com.pengxh.app.weatherplus.service.LocationService;
-import com.pengxh.app.weatherplus.ui.fragment.OtherWeatherFragment;
-import com.pengxh.app.weatherplus.ui.fragment.WeatherFragment;
 import com.pengxh.app.weatherplus.utils.SQLiteUtil;
+import com.pengxh.app.weatherplus.widgets.EasyPopupWindow;
 
 import java.util.ArrayList;
-import java.util.LinkedList;
+import java.util.Arrays;
 import java.util.List;
 
 import butterknife.BindView;
+import butterknife.OnClick;
 
-public class MainActivity extends BaseNormalActivity {
+public class MainActivity extends BaseNormalActivity implements View.OnClickListener {
 
     private static final String TAG = "MainActivity";
 
+    @BindView(R.id.layoutView)
+    LinearLayout layoutView;
     @BindView(R.id.mMainViewPager)
     ViewPager mMainViewPager;
     @BindView(R.id.mLlIndicator)
     LinearLayout mLlIndicator;
+    private List<String> items = Arrays.asList("管理城市", "更新间隔");
+    private Intent locationIntent = null;
+    private String currentLocation = "";
     private SQLiteUtil sqLiteUtil;
     private PageNumberUpdateBroadcast updateBroadcast = null;
 
     @Override
-    public void initView() {
-        setContentView(R.layout.activity_main);
+    public int initLayoutView() {
+        return R.layout.activity_main;
     }
 
     @Override
     public void initData() {
+        ImmersionBar.with(this).statusBarColor(R.color.statusBar_color).fitsSystemWindows(true).init();
+        locationIntent = new Intent(this, LocationService.class);
+        startService(locationIntent);
         sqLiteUtil = SQLiteUtil.getInstance();
     }
 
     @Override
     public void initEvent() {
-        mHandler.sendEmptyMessage(20);
+        currentLocation = (String) SaveKeyValues.getValue("location", "");
+    }
+
+    @OnClick(R.id.manageCity)
+    @Override
+    public void onClick(View v) {
+        EasyPopupWindow easyPopupWindow = new EasyPopupWindow(this, items);
+        easyPopupWindow.setPopupWindowClickListener(position -> {
+            if (position == 0) {
+                Intent intent = new Intent(this, SelectCityActivity.class);
+                intent.putExtra("currentLocation", currentLocation);
+                startActivity(intent);
+            } else if (position == 1) {
+
+            }
+        });
+        easyPopupWindow.setBackgroundDrawable(null);
+        easyPopupWindow.showAsDropDown(layoutView,
+                layoutView.getWidth() - easyPopupWindow.getWidth() - DensityUtil.dp2px(this, 15),
+                DensityUtil.dp2px(this, 40));
+    }
+
+    @Override
+    protected void onDestroy() {
+        super.onDestroy();
+        if (updateBroadcast != null) {
+            unregisterReceiver(updateBroadcast);
+        }
+        if (locationIntent == null) {
+            return;
+        }
+        stopService(locationIntent);
     }
 
     class PageNumberUpdateBroadcast extends BroadcastReceiver {
@@ -65,109 +98,19 @@ public class MainActivity extends BaseNormalActivity {
         public void onReceive(Context context, Intent intent) {
             String action = intent.getAction();
             if (action != null && action.equals("action.updatePageNumber")) {
-                String tag = intent.getStringExtra("TAG");
-                Log.d(TAG, "onReceive: => " + tag);
-                switch (tag) {
-                    case "add":
-                        sengMsg(sqLiteUtil.loadCityList().size());
-                        break;
-                    case "del":
-                        sengMsg(sqLiteUtil.loadCityList().size());
-                        break;
-                    default:
-                        Log.w(TAG, "onReceive: error tag", new Throwable());
-                        break;
-                }
+
             }
         }
     }
 
-    private void sengMsg(int number) {
-        Message message = mHandler.obtainMessage();
-        message.what = 30;
-        message.obj = number;
-        mHandler.sendMessage(message);
-    }
-
-    @SuppressLint("HandlerLeak")
-    private Handler mHandler = new Handler() {
-        @Override
-        public void handleMessage(Message msg) {
-            switch (msg.what) {
-                case 30:
-                    //刷新UI
-                    int number = (int) msg.obj;
-                    updatePage(number, true);
-                    break;
-                default:
-                    //默认加载
-                    updatePage(sqLiteUtil.loadCityList().size(), false);
-                    break;
-            }
-        }
-    };
-
-    private void updatePage(int pageNum, boolean isUpdateNumer) {
-        Log.d(TAG, "isUpdate => " + isUpdateNumer);
-        List<Fragment> fragments = new LinkedList<>();
-        if (!isUpdateNumer) {
-            fragments.add(new WeatherFragment());
-            for (int i = 1; i <= pageNum; i++) {
-                fragments.add(new OtherWeatherFragment());
-            }
-            WeatherPageAdapter pageAdapter = new WeatherPageAdapter(getSupportFragmentManager(), fragments);
-            mMainViewPager.setOffscreenPageLimit(2 * pageNum + 1);
-            mMainViewPager.setAdapter(pageAdapter);
-            mMainViewPager.setOnPageChangeListener(
-                    //page+1是因为需要将定位点页面计算在内
-                    new WeatherPageChangeListener(this, mLlIndicator, (pageNum + 1)));
-        } else {
-            Log.d(TAG, "pageNumber => " + (pageNum + 1));
-        }
-    }
-
-    @Override
-    protected void onStart() {
-        super.onStart();
-        //防止定位不准确
-        int p1 = ContextCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION);
-        int p2 = ContextCompat.checkSelfPermission(this, Manifest.permission.ACCESS_COARSE_LOCATION);
-        int p3 = ContextCompat.checkSelfPermission(this, Manifest.permission.READ_PHONE_STATE);
-        if (PackageManager.PERMISSION_GRANTED != p1 || PackageManager.PERMISSION_GRANTED != p2 || PackageManager.PERMISSION_GRANTED != p3) {
-            new AlertView("友情提示", "缺少定位权限，请在设置里面打开相关权限", null, new String[]{"确定"}, null, MainActivity.this,
-                    AlertView.Style.Alert, new OnItemClickListener() {
-                @Override
-                public void onItemClick(Object o, int position) {
-                    MainActivity.this.finish();
-                }
-            }).show();
-        } else {
-            startService(new Intent(this, LocationService.class));
-        }
-        if (updateBroadcast == null) {
-            updateBroadcast = new PageNumberUpdateBroadcast();
-            IntentFilter intentFilter = new IntentFilter();
-            intentFilter.addAction("action.updatePageNumber");
-            this.registerReceiver(updateBroadcast, intentFilter);
-        }
-    }
-
-    @Override
-    protected void onDestroy() {
-        super.onDestroy();
-        if (updateBroadcast != null) {
-            this.unregisterReceiver(updateBroadcast);
-        }
-    }
-
-    class WeatherPageChangeListener implements ViewPager.OnPageChangeListener {
+    private class WeatherPageChangeListener implements ViewPager.OnPageChangeListener {
 
         private int mPageCount;//页数
         private List<ImageView> mImgList;//保存img总个数
         private int img_select;
         private int img_unSelect;
 
-        WeatherPageChangeListener(Context context, LinearLayout mLlIndicator, int pageCount) {
+        private WeatherPageChangeListener(Context context, LinearLayout mLlIndicator, int pageCount) {
             this.mPageCount = pageCount;
 
             mImgList = new ArrayList<>();
