@@ -4,39 +4,29 @@ import android.annotation.SuppressLint;
 import android.app.ProgressDialog;
 import android.os.Handler;
 import android.os.Message;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.StaggeredGridLayoutManager;
-import android.text.Editable;
 import android.text.TextUtils;
-import android.text.TextWatcher;
+import android.util.Log;
 import android.view.View;
 import android.widget.ArrayAdapter;
 import android.widget.AutoCompleteTextView;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.recyclerview.widget.StaggeredGridLayoutManager;
+
 import com.aihook.alertview.library.AlertView;
-import com.aihook.alertview.library.OnItemClickListener;
 import com.alibaba.fastjson.JSONObject;
 import com.gyf.immersionbar.ImmersionBar;
 import com.pengxh.app.multilib.base.BaseNormalActivity;
 import com.pengxh.app.weatherplus.R;
 import com.pengxh.app.weatherplus.adapter.HotCityAdapter;
-import com.pengxh.app.weatherplus.bean.AllCityBean;
-import com.pengxh.app.weatherplus.bean.CityNameBean;
-import com.pengxh.app.weatherplus.bean.HotCityNameBean;
+import com.pengxh.app.weatherplus.bean.CityInfoBean;
+import com.pengxh.app.weatherplus.bean.HotCityBean;
 import com.pengxh.app.weatherplus.bean.NetWeatherBean;
-import com.pengxh.app.weatherplus.event.AutoCompleteEvent;
-import com.pengxh.app.weatherplus.event.TagEvent;
 import com.pengxh.app.weatherplus.mvp.presenter.WeatherPresenterImpl;
 import com.pengxh.app.weatherplus.mvp.view.IWeatherView;
-import com.pengxh.app.weatherplus.utils.GreenDaoUtil;
-import com.pengxh.app.weatherplus.utils.OtherUtil;
 import com.pengxh.app.weatherplus.utils.SQLiteUtil;
-
-import org.greenrobot.eventbus.EventBus;
-import org.greenrobot.eventbus.Subscribe;
-import org.greenrobot.eventbus.ThreadMode;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -44,8 +34,13 @@ import java.util.List;
 import butterknife.BindView;
 import butterknife.OnClick;
 
-public class SelectCityActivity extends BaseNormalActivity implements IWeatherView, View.OnClickListener, OnItemClickListener {
+public class SelectCityActivity extends BaseNormalActivity implements IWeatherView, View.OnClickListener {
 
+    private static final String TAG = "SelectCityActivity";
+    @BindView(R.id.mTextView_title)
+    TextView mTextView_title;
+    @BindView(R.id.mImageView_title_add)
+    ImageView mImageView_title_add;
     @BindView(R.id.mTextView_current_location)
     TextView mTextView_current_location;
     @BindView(R.id.mImageView_hot_city)
@@ -54,165 +49,133 @@ public class SelectCityActivity extends BaseNormalActivity implements IWeatherVi
     AutoCompleteTextView mAutoCompleteTextView;
     @BindView(R.id.mRecyclerView_hot_city)
     RecyclerView mRecyclerViewHotCity;
-    @BindView(R.id.mImageView_title_add)
-    ImageView mImageView_title_add;
-    @BindView(R.id.mTextView_title)
-    TextView mTextView_title;
 
+    private SQLiteUtil sqLiteUtil;
+    private List<HotCityBean> hotCityList = new ArrayList<>();
+    private HotCityAdapter hotCityAdapter = null;
     private WeatherPresenterImpl weatherPresenter;
     private ProgressDialog progressDialog;
-    private HotCityAdapter hotCityAdapter;
     private AlertView alertView;
-    private List<HotCityNameBean> allHotCity;
-    private SQLiteUtil sqLiteUtil;
 
     @Override
-    public void initView() {
-        setContentView(R.layout.activity_selectcity);
-        ImmersionBar.with(this)
-                .statusBarColor("#0094FF")
-                .fitsSystemWindows(true)
-                .init();
+    public int initLayoutView() {
+        return R.layout.activity_selectcity;
     }
 
     @Override
     public void initData() {
+        ImmersionBar.with(this).statusBarColor("#0094FF").fitsSystemWindows(true).init();
         mImageView_title_add.setVisibility(View.INVISIBLE);
         mTextView_title.setText("添加城市");
-        String district = getIntent().getStringExtra("district");
+
+        String district = getIntent().getStringExtra("currentLocation");
         if (!TextUtils.isEmpty(district)) {
             mTextView_current_location.setText(district);
         } else {
             mTextView_current_location.setText("定位失败");
         }
+
+        weatherPresenter = new WeatherPresenterImpl(this);
+        sqLiteUtil = SQLiteUtil.getInstance();
+        //加载所有城市名并转为数组
+        new Thread(() -> {
+            String[] array = sqLiteUtil.loadAllCityName();
+            //获取到数组之后再去初始化AutoCompleteTextView的Adapter
+            Message message = mHandler.obtainMessage();
+            message.what = 1000;
+            message.obj = array;
+            mHandler.sendMessage(message);
+        }).start();
     }
 
     @Override
     public void initEvent() {
-        sqLiteUtil = SQLiteUtil.getInstance();
-        weatherPresenter = new WeatherPresenterImpl(this);
-
-        new Thread(new Runnable() {
-            @Override
-            public void run() {
-                List<CityNameBean> allCityName = GreenDaoUtil.loadAllCityName();
-                List<String> cities = new ArrayList<>();
-                for (int i = 0; i < allCityName.size(); i++) {
-                    String city = allCityName.get(i).getCity();
-                    cities.add(city);
-                }
-                EventBus.getDefault().postSticky(new AutoCompleteEvent(cities));
-            }
-        }).start();
-
-        allHotCity = sqLiteUtil.loadHotCity();
-        if (allHotCity.size() > 0) {
+        hotCityList = sqLiteUtil.loadHotCity();
+        if (hotCityList != null) {
             mImageView_hot_city.setVisibility(View.VISIBLE);
-            hotCityAdapter = new HotCityAdapter(this, allHotCity);
-            mRecyclerViewHotCity.setLayoutManager(
-                    new StaggeredGridLayoutManager(4, StaggeredGridLayoutManager.VERTICAL));
-            mRecyclerViewHotCity.setAdapter(hotCityAdapter);
-            hotCityAdapter.setOnItemClickListener(new HotCityAdapter.OnItemClickListener() {
-                @Override
-                public void onClick(int position) {
-                    String city = allHotCity.get(position).getCityName();
-
-                    List<AllCityBean> beanList = GreenDaoUtil.queryCity(city);
-                    if (beanList.size() > 0) {
-                        getCityWeather(beanList.get(0));
-                    }
-                }
-            });
         } else {
             mImageView_hot_city.setVisibility(View.INVISIBLE);
         }
-    }
-
-    @Subscribe(sticky = true, threadMode = ThreadMode.MAIN)
-    public void onEventMainThread(AutoCompleteEvent event) {
-        List<String> cities = event.getCities();
-        ArrayAdapter<String> adapter = new ArrayAdapter<>(this
-                , android.R.layout.simple_dropdown_item_1line
-                , OtherUtil.removeDuplicate(cities));
-        mAutoCompleteTextView.setThreshold(1); //设置输入一个字符提示，默认为2
-        mAutoCompleteTextView.setAdapter(adapter);
-        mAutoCompleteTextView.addTextChangedListener(new TextWatcher() {
+        hotCityAdapter = new HotCityAdapter(SelectCityActivity.this, hotCityList);
+        mRecyclerViewHotCity.setLayoutManager(new StaggeredGridLayoutManager(4, StaggeredGridLayoutManager.VERTICAL));
+        mRecyclerViewHotCity.setAdapter(hotCityAdapter);
+        hotCityAdapter.setOnItemClickListener(new HotCityAdapter.OnItemClickListener() {
             @Override
-            public void beforeTextChanged(CharSequence s, int start, int count, int after) {
-
-            }
-
-            @Override
-            public void onTextChanged(CharSequence s, int start, int before, int count) {
-
-            }
-
-            @Override
-            public void afterTextChanged(Editable s) {
-                String inputString = s.toString();
-                List<AllCityBean> beanList = GreenDaoUtil.queryCity(inputString);
-                if (beanList.size() > 0) {
-                    //将查询历史保存到[热门]表
-                    AllCityBean cityBean = beanList.get(0);
-
-                    sqLiteUtil.saveHotCity(cityBean.getCity());
+            public void onClick(int position) {
+                String city = hotCityList.get(position).getCity();
+                CityInfoBean.ResultBeanX.ResultBean cityBean = sqLiteUtil.queryCityInfo(city);
+                if (cityBean != null) {
                     getCityWeather(cityBean);
                 }
             }
         });
-        EventBus.getDefault().removeStickyEvent(event);
     }
 
-    private void getCityWeather(AllCityBean cityBean) {
-        weatherPresenter.onReadyRetrofitRequest(cityBean.getCity(),
-                Integer.parseInt(cityBean.getCityid()),
-                Integer.parseInt(cityBean.getCitycode()));
-        //通知CityListActivity刷新UI
-        EventBus.getDefault().postSticky(new TagEvent(SelectCityActivity.class.getSimpleName(), 1));
+    private void getCityWeather(CityInfoBean.ResultBeanX.ResultBean cityBean) {
+        Log.d(TAG, "getCityWeather: 查询天气");
+        weatherPresenter.onReadyRetrofitRequest(cityBean.getCity(), cityBean.getCityid(), Integer.parseInt(cityBean.getCitycode()));
+//        EventBus.getDefault().postSticky(new TagEvent(SelectCityActivity.class.getSimpleName(), 1));
     }
+
+    @SuppressLint("HandlerLeak")
+    private Handler mHandler = new Handler() {
+        @Override
+        public void handleMessage(Message msg) {
+            switch (msg.what) {
+                case 10:
+                    finish();
+                    break;
+                case 1000:
+                    String[] cityNameArray = (String[]) msg.obj;
+                    ArrayAdapter<String> adapter = new ArrayAdapter<>(SelectCityActivity.this, android.R.layout.simple_dropdown_item_1line, cityNameArray);
+                    mAutoCompleteTextView.setThreshold(1); //设置输入一个字符提示，默认为2
+                    mAutoCompleteTextView.setAdapter(adapter);
+                    mAutoCompleteTextView.setOnItemClickListener((parent, view, position, id) -> {
+                        String text = mAutoCompleteTextView.getText().toString();
+                        CityInfoBean.ResultBeanX.ResultBean cityBean = sqLiteUtil.queryCityInfo(text);
+                        if (cityBean != null) {
+                            String city = cityBean.getCity();
+                            //将查询历史保存到[热门]表
+                            sqLiteUtil.saveHotCity(city);
+                            //查询天气
+                            getCityWeather(cityBean);
+
+                            for (HotCityBean bean : hotCityList) {
+                                if (city.equals(bean.getCity())) {
+                                    Log.d(TAG, "热门城市已存在，不更新列表");
+                                    return;
+                                }
+                            }
+                            HotCityBean hotCityNameBean = new HotCityBean();
+                            hotCityNameBean.setCity(city);
+                            hotCityList.add(hotCityNameBean);
+                            hotCityAdapter.notifyDataSetChanged();
+                            mImageView_hot_city.setVisibility(View.VISIBLE);
+                        }
+                    });
+                    break;
+                default:
+                    throw new IllegalStateException("Unexpected value: " + msg.what);
+            }
+        }
+    };
 
     @OnClick({R.id.mImageView_title_back, R.id.mImageView_hot_city})
     @Override
     public void onClick(View v) {
         switch (v.getId()) {
             case R.id.mImageView_title_back:
-                runOnUiThread(new Runnable() {
-                    @Override
-                    public void run() {
-                        //在主线程销毁，直接出栈，返回键不会返回到前一个页面
-                        finish();
-                    }
-                });
+                finish();
                 break;
             case R.id.mImageView_hot_city:
-                alertView = new AlertView(
-                        "提示",
-                        "确认删除全部搜索历史?",
-                        "取消",
-                        new String[]{"确定"},
-                        null,
-                        this,
-                        AlertView.Style.Alert,
-                        this).setCancelable(false);
-                alertView.show();
-                break;
-        }
-    }
-
-    @Override
-    public void onItemClick(Object o, int position) {
-        /**
-         * -1取消，0确定
-         * */
-        switch (position) {
-            case -1:
-                alertView.dismiss();
-                break;
-            case 0:
-                allHotCity.clear();
-                sqLiteUtil.deleteAll();
-                hotCityAdapter.notifyDataSetChanged();
-                mImageView_hot_city.setVisibility(View.INVISIBLE);
+                new AlertView("提示", "确认删除全部搜索历史?", "取消", new String[]{"确定"}, null, this, AlertView.Style.Alert, (o, position) -> {
+                    if (position == 0) {
+                        hotCityList.clear();
+                        sqLiteUtil.deleteAll();
+                        hotCityAdapter.notifyDataSetChanged();
+                        mImageView_hot_city.setVisibility(View.INVISIBLE);
+                    }
+                }).setCancelable(false).show();
                 break;
         }
     }
@@ -232,19 +195,9 @@ public class SelectCityActivity extends BaseNormalActivity implements IWeatherVi
         if (progressDialog.isShowing()) {
             progressDialog.dismiss();
             //延时关闭页面，确保CityListActivity页面能基本收到消息
-            mHandler.sendEmptyMessageDelayed(10, 500);
+//            mHandler.sendEmptyMessageDelayed(10, 500);
         }
     }
-
-    @SuppressLint("HandlerLeak")
-    private Handler mHandler = new Handler() {
-        @Override
-        public void handleMessage(Message msg) {
-            if (msg.what == 10) {
-                finish();
-            }
-        }
-    };
 
     @Override
     public void showNetWorkData(NetWeatherBean weatherBean) {
@@ -254,18 +207,6 @@ public class SelectCityActivity extends BaseNormalActivity implements IWeatherVi
             String jsonString = JSONObject.toJSONString(weatherBean);
             sqLiteUtil.saveCityListWeather(city, jsonString);
         }
-    }
-
-    @Override
-    protected void onStart() {
-        super.onStart();
-        EventBus.getDefault().register(this);
-    }
-
-    @Override
-    protected void onStop() {
-        EventBus.getDefault().unregister(this);
-        super.onStop();
     }
 
     @Override
